+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to the **daily-python** SDK will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- Added support to automatically start a recording when joining a room if the `start_cloud_recording` token property is
+  set to `True`.
+
 ## [0.14.2] - 2024-12-22
 
 ### Fixed